requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.0  # Ajouté pour les appels HTTP vers Azure OpenAI